+++ minder_mobile/lib/main.dart
@@ -2,7 +2,7 @@ import 'package:animated_splash_screen/animated_splash_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
-import 'screens/sign_up.dart';
+import 'screens/sign_in.dart';
 
 void main() {
   runApp(const MyApp());
